[o11y AI] Refactor tools (#247070)
Refactors the o11y tools to a consistent separation between tool
definition and handling logic. Previously, tools mixed these and used
inconsistent naming conventions, making it harder to navigate (both for
humans and LLMs)

New folder structure
-  `tool.ts`: Contains the tool schema, metadata  and registration
- `handler.ts`: Encapsulates the business logic (mostly data retrieval
and data formatting)

**Advantages**
- This improves maintainability and readability.
- This will enable us to consume tools from AI insight routes by calling
the tool handler directly (bypassing the tool wrapper)

(cherry picked from commit 101cdda3fa92000a95b02b3ccb97a5455b0d3a78)
