+++ x-pack/solutions/observability/plugins/observability_agent_builder/server/tools/get_log_categories/tool.ts
@@ -1,0 +1,127 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { z } from '@kbn/zod';
+import { ToolType } from '@kbn/onechat-common';
+import { ToolResultType } from '@kbn/onechat-common/tools/tool_result';
+import type { BuiltinToolDefinition, StaticToolRegistration } from '@kbn/onechat-server';
+import type { CoreSetup, Logger } from '@kbn/core/server';
+import type {
+  ObservabilityAgentBuilderPluginStart,
+  ObservabilityAgentBuilderPluginStartDependencies,
+} from '../../types';
+import { timeRangeSchemaOptional, indexDescription } from '../../utils/tool_schemas';
+import { getAgentBuilderResourceAvailability } from '../../utils/get_agent_builder_resource_availability';
+import type { getFilteredLogCategories } from './handler';
+import { getToolHandler } from './handler';
+
+export interface GetLogCategoriesToolResult {
+  type: ToolResultType.other;
+  data: {
+    highSeverityCategories: Awaited<ReturnType<typeof getFilteredLogCategories>>;
+    lowSeverityCategories: Awaited<ReturnType<typeof getFilteredLogCategories>>;
+  };
+}
+
+const DEFAULT_TIME_RANGE = {
+  start: 'now-1h',
+  end: 'now',
+};
+
+export const OBSERVABILITY_GET_LOG_CATEGORIES_TOOL_ID = 'observability.get_log_categories';
+
+const getLogsSchema = z.object({
+  ...timeRangeSchemaOptional(DEFAULT_TIME_RANGE),
+  index: z.string().describe(indexDescription).optional(),
+  terms: z
+    .record(z.string(), z.string())
+    .optional()
+    .describe(
+      'Optional field filters to narrow down results. Each key-value pair filters logs where the field exactly matches the value. Example: { "service.name": "payment", "host.name": "web-server-01" }. Multiple filters are combined with AND logic.'
+    ),
+});
+
+export function createGetLogCategoriesTool({
+  core,
+  logger,
+}: {
+  core: CoreSetup<
+    ObservabilityAgentBuilderPluginStartDependencies,
+    ObservabilityAgentBuilderPluginStart
+  >;
+  logger: Logger;
+}): StaticToolRegistration<typeof getLogsSchema> {
+  const toolDefinition: BuiltinToolDefinition<typeof getLogsSchema> = {
+    id: OBSERVABILITY_GET_LOG_CATEGORIES_TOOL_ID,
+    type: ToolType.builtin,
+    description: `Compresses thousands of logs into a small set of categories to provide a high-level overview of what's being logged.
+
+When to use:
+- Getting a quick summary of log activity in a service or time range
+- Identifying the types of events occurring without reading individual logs
+- Discovering unexpected log patterns or new error types
+- Answering "what kinds of things are happening?" rather than "what exactly happened?"
+
+How it works:
+Groups similar log messages together using pattern recognition, returning representative categories with counts.
+
+Do NOT use for:
+- Understanding the sequence of events for a specific error (use get_correlated_logs)
+- Investigating a specific incident in detail (use get_correlated_logs)
+- Analyzing changes in log volume over time (use run_log_rate_analysis)`,
+    schema: getLogsSchema,
+    tags: ['observability', 'logs'],
+    availability: {
+      cacheMode: 'space',
+      handler: async ({ request }) => {
+        return getAgentBuilderResourceAvailability({ core, request, logger });
+      },
+    },
+    handler: async (
+      { index, start = DEFAULT_TIME_RANGE.start, end = DEFAULT_TIME_RANGE.end, terms },
+      { esClient }
+    ) => {
+      try {
+        const { highSeverityCategories, lowSeverityCategories } = await getToolHandler({
+          core,
+          logger,
+          esClient,
+          index,
+          start,
+          end,
+          terms,
+        });
+
+        return {
+          results: [
+            {
+              type: ToolResultType.other,
+              data: { highSeverityCategories, lowSeverityCategories },
+            },
+          ],
+        };
+      } catch (error) {
+        logger.error(`Error fetching log categories: ${error.message}`);
+        logger.debug(error);
+
+        return {
+          results: [
+            {
+              type: ToolResultType.error,
+              data: {
+                message: `Failed to fetch log categories: ${error.message}`,
+                stack: error.stack,
+              },
+            },
+          ],
+        };
+      }
+    },
+  };
+
+  return toolDefinition;
+}
