+++ x-pack/solutions/observability/plugins/observability_agent_builder/server/tools/index.ts
@@ -5,12 +5,12 @@
  * 2.0.
  */
 
-export { OBSERVABILITY_GET_DATA_SOURCES_TOOL_ID } from './get_data_sources/get_data_sources';
-export { OBSERVABILITY_GET_ALERTS_TOOL_ID, defaultFields } from './get_alerts/get_alerts';
-export { OBSERVABILITY_GET_LOG_CATEGORIES_TOOL_ID } from './get_log_categories/get_log_categories';
-export { OBSERVABILITY_RUN_LOG_RATE_ANALYSIS_TOOL_ID } from './run_log_rate_analysis/run_log_rate_analysis';
-export { OBSERVABILITY_GET_ANOMALY_DETECTION_JOBS_TOOL_ID } from './get_anomaly_detection_jobs/get_anomaly_detection_jobs';
-export { OBSERVABILITY_GET_SERVICES_TOOL_ID } from './get_services/get_services';
-export { OBSERVABILITY_GET_DOWNSTREAM_DEPENDENCIES_TOOL_ID } from './get_downstream_dependencies/get_downstream_dependencies';
-export { OBSERVABILITY_GET_CORRELATED_LOGS_TOOL_ID } from './get_correlated_logs/get_correlated_logs';
-export { OBSERVABILITY_GET_HOSTS_TOOL_ID } from './get_hosts/get_hosts';
+export { OBSERVABILITY_GET_DATA_SOURCES_TOOL_ID } from './get_data_sources/tool';
+export { OBSERVABILITY_GET_ALERTS_TOOL_ID, defaultFields } from './get_alerts/tool';
+export { OBSERVABILITY_GET_LOG_CATEGORIES_TOOL_ID } from './get_log_categories/tool';
+export { OBSERVABILITY_RUN_LOG_RATE_ANALYSIS_TOOL_ID } from './run_log_rate_analysis/tool';
+export { OBSERVABILITY_GET_ANOMALY_DETECTION_JOBS_TOOL_ID } from './get_anomaly_detection_jobs/tool';
+export { OBSERVABILITY_GET_SERVICES_TOOL_ID } from './get_services/tool';
+export { OBSERVABILITY_GET_DOWNSTREAM_DEPENDENCIES_TOOL_ID } from './get_downstream_dependencies/tool';
+export { OBSERVABILITY_GET_CORRELATED_LOGS_TOOL_ID } from './get_correlated_logs/tool';
+export { OBSERVABILITY_GET_HOSTS_TOOL_ID } from './get_hosts/tool';
