+++ x-pack/solutions/observability/plugins/observability_agent_builder/server/tools/register_tools.ts
@@ -17,33 +17,30 @@ import type { ObservabilityAgentBuilderDataRegistry } from '../data_registry/dat
 import {
   OBSERVABILITY_GET_DATA_SOURCES_TOOL_ID,
   createGetDataSourcesTool,
-} from './get_data_sources/get_data_sources';
+} from './get_data_sources/tool';
 import {
   OBSERVABILITY_RUN_LOG_RATE_ANALYSIS_TOOL_ID,
   createRunLogRateAnalysisTool,
-} from './run_log_rate_analysis/run_log_rate_analysis';
+} from './run_log_rate_analysis/tool';
 import {
   OBSERVABILITY_GET_ANOMALY_DETECTION_JOBS_TOOL_ID,
   createGetAnomalyDetectionJobsTool,
-} from './get_anomaly_detection_jobs/get_anomaly_detection_jobs';
-import { OBSERVABILITY_GET_ALERTS_TOOL_ID, createGetAlertsTool } from './get_alerts/get_alerts';
+} from './get_anomaly_detection_jobs/tool';
+import { OBSERVABILITY_GET_ALERTS_TOOL_ID, createGetAlertsTool } from './get_alerts/tool';
 import {
   OBSERVABILITY_GET_LOG_CATEGORIES_TOOL_ID,
   createGetLogCategoriesTool,
-} from './get_log_categories/get_log_categories';
+} from './get_log_categories/tool';
 import {
   OBSERVABILITY_GET_CORRELATED_LOGS_TOOL_ID,
   createGetCorrelatedLogsTool,
-} from './get_correlated_logs/get_correlated_logs';
-import { OBSERVABILITY_GET_HOSTS_TOOL_ID, createGetHostsTool } from './get_hosts/get_hosts';
-import {
-  createGetServicesTool,
-  OBSERVABILITY_GET_SERVICES_TOOL_ID,
-} from './get_services/get_services';
+} from './get_correlated_logs/tool';
+import { OBSERVABILITY_GET_HOSTS_TOOL_ID, createGetHostsTool } from './get_hosts/tool';
+import { createGetServicesTool, OBSERVABILITY_GET_SERVICES_TOOL_ID } from './get_services/tool';
 import {
   createDownstreamDependenciesTool,
   OBSERVABILITY_GET_DOWNSTREAM_DEPENDENCIES_TOOL_ID,
-} from './get_downstream_dependencies/get_downstream_dependencies';
+} from './get_downstream_dependencies/tool';
 
 const PLATFORM_TOOL_IDS = [
   platformCoreTools.search,
