+++ x-pack/solutions/observability/test/api_integration_deployment_agnostic/apis/observability_agent_builder/tools/get_anomaly_detection_jobs.spec.ts
@@ -7,8 +7,8 @@
 
 import expect from '@kbn/expect';
 import type { ApmSynthtraceEsClient } from '@kbn/synthtrace';
-import type { GetAnomalyDetectionJobsToolResult } from '@kbn/observability-agent-builder-plugin/server/tools/get_anomaly_detection_jobs/get_anomaly_detection_jobs';
-import { OBSERVABILITY_GET_ANOMALY_DETECTION_JOBS_TOOL_ID } from '@kbn/observability-agent-builder-plugin/server/tools/get_anomaly_detection_jobs/get_anomaly_detection_jobs';
+import type { GetAnomalyDetectionJobsToolResult } from '@kbn/observability-agent-builder-plugin/server/tools/get_anomaly_detection_jobs/tool';
+import { OBSERVABILITY_GET_ANOMALY_DETECTION_JOBS_TOOL_ID } from '@kbn/observability-agent-builder-plugin/server/tools/get_anomaly_detection_jobs/tool';
 import datemath from '@elastic/datemath';
 import moment from 'moment';
 import type { DeploymentAgnosticFtrProviderContext } from '../../../ftr_provider_context';
